Extract functions and use labels over individual issues
